+++ example/lib/widgets/event_calendar/home_page.dart
@@ -1,11 +1,8 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_event_calendar/flutter_event_calendar.dart';
-import 'package:flutter_event_calendar_example/models/calendar_event_model.dart';
+import 'package:flutter_event_calendar_example/widgets/event_calendar/eventCalendar.dart';
 import 'package:flutter_event_calendar_example/widgets/setting/setting.dart';
 import 'package:flutter_event_calendar_example/widgets/code_preview/code_preview.dart';
-import '../../injection.dart';
+
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({
@@ -17,20 +14,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  StreamController<CalendarEventModel> streamController =
-      getit<StreamController<CalendarEventModel>>();
-  late Stream stream;
-  late StreamSubscription<CalendarEventModel> streamSubscription;
-  CalendarEventModel calendarEventModel = getit<CalendarEventModel>();
-  @override
-  void initState() {
-    streamController.stream.listen((event) {
-      setState(() {
-        calendarEventModel = event;
-      });
-    });
-    super.initState();
-  }
+
 
   @override
   Widget build(BuildContext context) {
@@ -52,29 +36,7 @@ class _MyHomePageState extends State<MyHomePage> {
         ),
         body: TabBarView(
             children: [
-          EventCalendar(
-            specialDays:calendarEventModel.specialDays,
-            showLoadingForEvent: true,
-            calendarType: calendarEventModel.calendarType,
-            calendarLanguage: calendarEventModel.calendarLanguage,
-            calendarOptions: CalendarOptions(
-                viewType: calendarEventModel.calendarOptions.viewType,
-                headerMonthBackColor:
-                  calendarEventModel.calendarOptions.headerMonthBackColor,
-                toggleViewType:
-                    calendarEventModel.calendarOptions.toggleViewType,
-                font: calendarEventModel.calendarOptions.font),
-            dayOptions: DayOptions(
-                selectedBackgroundColor:
-                    calendarEventModel.dayOptions.selectedBackgroundColor),
-            headerOptions: HeaderOptions(
-              headerTextColor: calendarEventModel.headerOptions.headerTextColor,
-              navigationColor: calendarEventModel.headerOptions.navigationColor,
-              monthStringType: calendarEventModel.headerOptions.monthStringType,
-              weekDayStringType:
-                  calendarEventModel.headerOptions.weekDayStringType,
-            ),
-          ),
+          EventCalendarPage(),
           CodePreview(),
           Setting()
         ]),
